periodicity moved to new module
